Added support for irregular data grids
Function now allows plotting of irreguar data grids (where data is gridded
but not necessarily on a rectilinear grid).  There are a few edge cases
where the plotted results are a little wonky (e.g. NaNs on the edge of the
irregular-grid domain), but I think I'll wait to see if some of the bigger
issues get worked out in the HG2 contour objects (such as the resistance to
recoloring) before I look into that too closely.