Updated input to allow manual passing of private functions
In fast mode, the code calls a few different functions that are
typically hidden in the private folder in the Mapping Toolbox.
I've now added the ability for users to pass in these function
locations as parameters, in case the hard-coded default locations
aren't appropriate for various operating systems and Matlab releases.