New method for dealing with NaNs in calccontour
Implementing a more robust way of dealing with NaNs.  The previous
method worked decently for cartesian grids with only a few scattered
NaNs, but fell apart for non-cartesian grids and grid with unenclosed
NaN regions.

This new method should be robust to all types of NaN regions in all
structured (but not necessarily cartesian) grids.  So far, it works
decently, though polybool is throwing the occasional "polygon may
be invalid" for no apparent reason.  That's a multiplepolyint
issue, so going to try to fix it there...